Moved detectLanguage func to localization
Since localization will be needed in utils and this fixes circular
dependency
