+++ lib/utils.js
@@ -1,6 +1,5 @@
 'use strict'
 
-var qs = require('querystring')
 
 var L = window.L
 var moment = window.moment
@@ -132,95 +131,3 @@ function decodePolyline (polyline) {
     }
 
     module.exports.distanceString = distanceString
-
-    // Cookies functions from http://www.quirksmode.org/js/cookies.html
-    var createCookie = function (name, value, days) {
-      var expires = ''
-      if (days) {
-        var date = new Date()
-        date.setTime(date.getTime() + (days * 24 * 60 * 60 * 1000))
-        expires = '; expires=' + date.toGMTString()
-      }
-      document.cookie = name + '=' + value + expires + '; path=/'
-    }
-
-    var readCookie = function (name) {
-      var nameEQ = name + '='
-      var ca = document.cookie.split(';')
-      for (var i = 0;i < ca.length;i++) {
-        var c = ca[i]
-        while (c.charAt(0) == ' ') c = c.substring(1, c.length)
-        if (c.indexOf(nameEQ) == 0) return c.substring(nameEQ.length, c.length)
-      }
-      return null
-    }
-
-    var eraseCookie = function (name) {
-      createCookie(name, '', -1)
-    }
-
-    var detectLanguage = function () {
-      var detectLanguage = undefined
-
-      if (typeof window !== 'undefined') {
-        var qsParams = qs.parse(window.location.search)
-        console.log('Reading language from params:', window.location.search, qsParams)
-        if (qsParams[window.OTP_config.langQS]) {
-          detectLanguage = qsParams[window.OTP_config.langQS]
-          console.log('Param language:', detectLanguage)
-        }
-      }
-
-      // User wanted language isn't among languages we support
-      if (!_.has(window.OTP_config.availible_languages, detectLanguage)) {
-        detectLanguage = undefined
-      }
-
-      if (!detectLanguage && typeof document !== 'undefined') {
-        console.log('Reading language from cookie')
-        var cookieValue = readCookie(window.OTP_config.cookieName)
-        if (cookieValue) {
-          console.log('Got cookie: ', cookieValue)
-          detectLanguage = cookieValue
-          // User wanted language isn't among languages we support
-          if (!_.has(window.OTP_config.availible_languages, detectLanguage)) {
-            detectLanguage = undefined
-            eraseCookie(window.OTP_config.cookieName)
-          }
-        }
-      }
-
-      if (!detectLanguage && typeof navigator !== 'undefined') {
-        // Chooses among the wanted user language according to user preferences
-        if (navigator.languages) {
-          console.log('Navigator languages:', navigator.languages)
-          // returns first language that user wants that is availible
-          detectLanguage = _.find(navigator.languages, function getLang (ilang) {
-            console.log('currentLang', ilang, window.OTP_config.availible_languages)
-            return _.has(window.OTP_config.availible_languages, ilang)
-          })
-        } else {
-          detectLanguage = (navigator.language || navigator.userLanguage)
-        }
-      }
-      console.log('DetectedLanguage:', detectLanguage)
-      // Language wasn't found by any means - usually shouldn't happen
-      if (detectLanguage === undefined) {
-        console.log('Setting default language')
-        return 'en'
-      // Language was found but we don't necessary have the translation
-      } else {
-        // User wanted language isn't among languages we support
-        if (!_.has(window.OTP_config.availible_languages, detectLanguage)) {
-          console.log("Wanted language isn't found:", detectLanguage)
-          detectLanguage = 'en'
-          return detectLanguage
-        }
-        console.log('Saving wanted language in a cookie:', detectLanguage)
-        // Save wanted language in a cookie for 30 days
-        createCookie(window.OTP_config.cookieName, detectLanguage, 30)
-      }
-      return detectLanguage
-    }
-
-    module.exports.detectLanguage = detectLanguage
